+++ app/_utils/mixins.js
@@ -1,9 +1,9 @@
-import {css} from 'styled-components';
+import { css } from 'styled-components';
 
 export const media = {
-  desktop: (...args) => css`
-    @media (min-width: 1024px) {
-      ${ css(...args) }
-    }
-  `
-}
+    desktop: (...args) => css`
+        @media (min-width: 1024px) {
+            ${ css(...args) }
+        }
+    `
+};