+++ app/containers/Main/Main.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react';
 import TodoApp from 'components/TodoApp/TodoApp';
 import Heading from 'components/Heading/Heading';
 import TodoList from 'components/TodoList/TodoList';
@@ -16,8 +16,8 @@ const todos = [
     {
         task: 'Make dinner',
         isCompleted: false
-    }    
-]
+    }
+];
 
 export default class Main extends Component {
     constructor() {
@@ -28,24 +28,24 @@ export default class Main extends Component {
 
         this.state = {
             todos
-        }
+        };
     }
     addTodo(task) {
         const newState = [...this.state.todos];
         newState.unshift({
             task,
-            isCompleted: false        
+            isCompleted: false
         });
-        this.setState({todos: newState});
+        this.setState({ todos: newState });
     }
     removeTodo(id) {
-        const todos = this.state.todos;
+        const { todos } = this.state;
         const newState = [
             ...todos.slice(0, id),
             ...todos.slice(id + 1)
         ];
-        this.setState({todos: newState});
-    }    
+        this.setState({ todos: newState });
+    }
     render() {
         return (
             <TodoApp>
@@ -53,6 +53,6 @@ export default class Main extends Component {
                 <CreateTodo addTodo={this.addTodo}/>
                 <TodoList todos={this.state.todos} removeTodo={this.removeTodo}/>
             </TodoApp>
-        )
+        );
     }
 }