+++ app/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 
-import globalStyles from '_utils/global-styles';
+import _globalStyles from '_utils/global-styles';
 import Main from 'containers/Main/Main';
 
-render(<Main/>, document.getElementById('app'));
+render(<Main />, document.getElementById('app'));