+++ tools/webpack.config.prod.babel.js
@@ -1,6 +1,5 @@
-import {SRC,DIST,PORT} from './constants';
+import { SRC, DIST } from './constants';
 import webpack from 'webpack';
-import path from 'path';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 
 const HTMLWebpackPluginConfig = new HtmlWebpackPlugin({
@@ -15,28 +14,29 @@ const HTMLWebpackPluginConfig = new HtmlWebpackPlugin({
 
 module.exports = {
     entry: [
-        SRC + 'index.js' 
+        'babel-polyfill',
+        SRC + 'index.js'
     ],
     output: {
-        filename: "main.js",
+        filename: 'main.js',
         path: DIST,
-        publicPath: "/"
-    },    
+        publicPath: '/'
+    },
     module: {
         rules: [
             {
                 test: /\.js$/,
                 exclude: /node_modules/,
-                use: ['babel-loader']
+                use: ['babel-loader', 'eslint-loader']
             },
             {
-                test:   /\.css/,
+                test: /\.css/,
                 use: ['style-loader', 'css-loader']
-            },            
+            },
             {
                 test: /\.(jpe?g|png|gif|svg)$/i,
                 use: ['file-loader?name=images/[name].[ext]', 'img-loader?progressive=true']
-            },            
+            },
             {
                 test: /\.pug$/,
                 use: 'pug-loader?pretty=true'
@@ -65,4 +65,4 @@ module.exports = {
             debug: false
         })
     ]
-}
+};