+++ CHANGELOG.md
@@ -4,5 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2017-03-06
+### Added
+- eslint + airbnb config
+- babel polyfill -> polyfill Promises
+- babel env preset -> Default behavior without options is the same as babel-preset-latest
+
 ## [1.0.0] - 2017-01-09
 Initial release